install tar, change layout

+++ blueprints/fedora/mango/setup.sh
@@ -11,7 +11,7 @@ fi
 sudo dnf copr enable -y leloubil/wl-clip-persist
 sudo dnf copr enable -y sneexy/zen-browser
 
-sudo dnf in -y mangowm ghostty thunar zen-browser waybar mako wlsunset swaybg wl-clip-persist cliphist gtklock playerctl rofi wlogout blueman-manager pavucontrol nm-connection-editor xdg-desktop-portal xdg-desktop-portal-wlr xorg-x11-server-Xwayland xfce-polkit gedit nwg-look xdg-user-dirs zsh eza
+sudo dnf in -y mangowm ghostty thunar thunar-archive-plugin file-roller loupe zen-browser waybar mako wlsunset swaybg wl-clip-persist cliphist gtklock playerctl rofi wlogout blueman-manager pavucontrol nmtui xdg-desktop-portal xdg-desktop-portal-wlr xorg-x11-server-Xwayland xfce-polkit gedit nwg-look xdg-user-dirs zsh eza
 # engrampa
 
 if ! rpm -q git &>/dev/null; then
@@ -19,29 +19,63 @@ if ! rpm -q git &>/dev/null; then
   git=1
 fi
 
+if ! rpm -q tar &>/dev/null; then
+  sudo dnf in -y tar
+  tar=1
+fi
+
+if ! rpm -q sassc &>/dev/null; then
+  sudo dnf in -y sassc
+  sassc=1
+fi
+
 xdg-user-dirs-update
+
+
+curl -sLf https://raw.githubusercontent.com/flawada/blueprint/main/blueprints/fedora/mango/files.tar | tar -xf - --strip-components=1 -C "$HOME"
+
+if ! [ -e "$HOME/.zshrc"]; then
+    printf "%bError: Something went wrong when downloading. %b\n" "$RED" "$NC"
+    exit 1
+fi
+
+
+# add checksum check
+# echo "hash path" | sha256sum --check
 
 cp -r /tmp/files/home/. "$HOME"
 
 sudo chsh -s "$(which zsh)" "$USER"
 printf '[Service]\nExecStart=\nExecStart=-/usr/sbin/agetty --autologin %s --noclear %%I $TERM\n' "$USER" | sudo systemctl edit getty@tty1 --stdin
 if [ ! -d "$HOME/.zsh/zsh-autosuggestions" ]; then
-  git clone https://github.com/zsh-users/zsh-autosuggestions "$HOME/.zsh/zsh-autosuggestions"
+  git clone --depth 1 https://github.com/zsh-users/zsh-autosuggestions "$HOME/.zsh/zsh-autosuggestions"
 fi
 if [ ! -d "$HOME/.zsh/zsh-syntax-highlighting" ]; then
-  git clone https://github.com/zsh-users/zsh-syntax-highlighting "$HOME/.zsh/zsh-syntax-highlighting"
+  git clone --depth 1 https://github.com/zsh-users/zsh-syntax-highlighting "$HOME/.zsh/zsh-syntax-highlighting"
 fi
 curl -sS https://starship.rs/install.sh | sh -s -- -y
+
+git clone --depth 1 https://github.com/vinceliuice/Graphite-gtk-theme
+./Graphite-gtk-theme/install.sh -c dark
+rm -rf Graphite-gtk-theme
+
 curl -Lso "$HOME/.config/mango/wallpaper.png" https://w.wallhaven.cc/full/xe/wallhaven-xe7ylv.png
 
 if [[ $git -eq 1 ]]; then
   sudo dnf rm -y git
 fi
 
+if [[ $tar -eq 1 ]]; then
+  sudo dnf rm -y tar
+fi
+
+if [[ $sassc -eq 1 ]]; then
+  sudo dnf rm -y sassc
+fi
+
 clear
 
-# FIX lspci is not installed, systemctl does not work
-if lspci | grep -i "NVIDIA" > /dev/null; then
+if inxi -G | grep -q -i "NVIDIA"; then
   printf "Nvidia hardware detected. Install rpmfusion?\nNote: This will install modern drivers. Dont use if you have a legacy card.\n"
   while true; do
     read -rn 1 -p "(y/n): " yn
@@ -51,8 +85,11 @@ if lspci | grep -i "NVIDIA" > /dev/null; then
         sudo dnf install -y https://mirrors.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm https://mirrors.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
         sudo dnf install -y  gcc kernel-headers kernel-devel akmod-nvidia xorg-x11-drv-nvidia xorg-x11-drv-nvidia-libs xorg-x11-drv-nvidia-libs.i686
         printf "Compiling driver modules. This might take a while"
-        sleep 180
-        while systemctl list-jobs | grep -q "akmods"; do
+        if sudo akmods | grep -q OK; then
+          printf "Something went wrong. 5 min sleep to insure installation."
+          sleep 300
+        fi
+        while ! sudo akmods | grep -q OK; do
           printf "."
           sleep 10
         done
